refactor: extract forbidden usernames logic into its own file
Move FORBIDDEN_USERNAMES_REGEXPS, FORBIDDEN_REGEXP, and isForbiddenUsername
out of utilities.ts into a new forbidden_usernames.ts module so the
ever-growing list lives in a dedicated file.
